Fix hydration errors in Opted out task table
